Fix oath login

+++ app/src/main/java/de/max/roehrl/vueddit2/service/Util.kt
@@ -8,8 +8,8 @@ object Util {
         return "2 minutes ago"
     }
 
-    fun getUnixTime() : Int {
-        return System.currentTimeMillis().toInt()
+    fun getUnixTime() : Long {
+        return System.currentTimeMillis()
     }
 
     fun getAspectFixHeight(video: Video) : Int {
